+++ packages/desktop-app/src/renderer/shared/action-approval/manage/shortcuts/index.ts
@@ -1,0 +1,5 @@
+export { ShortcutsResetAllBindingsApproval } from "./ShortcutsResetAllBindingsApproval";
+export { ShortcutsResetBindingApproval } from "./ShortcutsResetBindingApproval";
+export { ShortcutsSetBindingApproval } from "./ShortcutsSetBindingApproval";
+export { ShortcutsSetQuickPanelBehaviorApproval } from "./ShortcutsSetQuickPanelBehaviorApproval";
+export { ShortcutsSetQuickPanelTriggerApproval } from "./ShortcutsSetQuickPanelTriggerApproval";